TIP-842: Move association type repository to structure sub context
