+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/repositories.yml
@@ -121,3 +121,16 @@ services:
             - '@pim_user.context.user'
             - '@doctrine.orm.entity_manager'
             - '%pim_catalog.entity.group_type.class%'
+
+    pim_enrich.repository.association_type:
+        class: 'Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Repository\InternalApi\AssociationTypeRepository'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.association_type.class%'
+
+    pim_catalog.repository.category:
+        class: 'Akeneo\Tool\Bundle\ClassificationBundle\Doctrine\ORM\Repository\CategoryRepository'
+        factory: 'doctrine.orm.entity_manager:getRepository'
+        arguments: ['%pim_catalog.entity.category.class%']
+        tags:
+            - { name: 'pim_repository' }
