+++ src/Pim/Bundle/CatalogBundle/Resources/config/repositories.yml
@@ -1,6 +1,5 @@
 parameters:
-    pim_catalog.repository.association_type.class:      Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\AssociationTypeRepository
-    pim_catalog.repository.category.class:              Akeneo\Tool\Bundle\ClassificationBundle\Doctrine\ORM\Repository\CategoryRepository
+    pim_catalog.repository.association_type.class:      Akeneo\Pim\Structure\Bundle\Doctrine\ORM\Repository\AssociationTypeRepository
     pim_catalog.repository.group.class:                 Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\GroupRepository
     pim_catalog.repository.product_unique_data.class:   Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\ProductUniqueDataRepository
     pim_catalog.repository.association.class:           Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\AssociationRepository
@@ -22,12 +21,6 @@ services:
         tags:
             - { name: 'pim_repository' }
 
-    pim_catalog.repository.category:
-        class: '%pim_catalog.repository.category.class%'
-        factory: 'doctrine.orm.entity_manager:getRepository'
-        arguments: ['%pim_catalog.entity.category.class%']
-        tags:
-            - { name: 'pim_repository' }
 
     pim_catalog.repository.group:
         class: '%pim_catalog.repository.group.class%'
